requirements.txt
cryptography==3.4.8
Flask==2.2.2
pysondb==1.6.7
python_bcrypt==0.3.2
